requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
joblib
scikit-learn